+++ src/metatrain/experimental/lorem/checkpoints.py
@@ -1,0 +1,5 @@
+"""Checkpoint upgrade helpers for the LOREM architecture.
+
+Version 1 is the initial experimental checkpoint format. Upgrade functions
+will be added here when the checkpoint schema changes.
+"""